feat(nodes, modal, storage): Add output nodes, modification modal and save functionality
In this commit, output nodes have been added to the application, expanding the node options available. A modal has also been implemented that allows users to modify the properties of these nodes. In addition, the functionality to save both edges and nodes to storage has been added, allowing the state of the application to be preserved between sessions.